Refactor duplicated query-protocol serialization code This eliminates the ec2 codegen entirely.  It's identical to the query protocol with 2 minor differences
